Moved UI classes from __init__.py to ui,py
Due to complexity UI related classes make, I moved every UI classes from __init__.py to ui.py. Having all UI related classes in ui.py will make things more simple.
